requirements.txt
azure-cosmos
azure-identity
Flask